# docs/swagger/swagger.yaml
...
info:
  contact: {}
  description: Payment service API (Gin)
  title: Modami Payment Service
  version: "1.0"
paths:
...
  /payments/return/momo:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: MoMo return URL
      tags:
      - Webhooks
  /payments/return/vnpay:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: VNPay return URL
      tags:
      - Webhooks
  /payments/return/zalopay:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: ZaloPay return URL
      tags:
      - Webhooks
...
  /webhooks/momo:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      parameters:
      - description: MoMo callback payload
        in: body
        name: payload
        schema:
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: MoMo IPN callback
      tags:
      - Webhooks
  /webhooks/vnpay:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      parameters:
      - description: VNPay callback payload (form or JSON)
        in: body
        name: payload
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: VNPay format response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: VNPay IPN callback
      tags:
      - Webhooks
  /webhooks/zalopay:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      parameters:
      - description: ZaloPay callback payload
        in: body
        name: payload
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: ZaloPay callback
      tags:
      - Webhooks
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"